refactor: rename the src/components folder to src/component-library
