+++ src/component-library/Layout/NavigationBar.style.css
@@ -14,3 +14,8 @@
   max-width: 100%;
   width: 200px;
 }
+
+.main-nav-item-container {
+  display: flex;
+  align-items: center;
+}
