+++ src/features/users/routes/Profile.jsx
@@ -1,7 +1,8 @@
 import { Link } from "react-router-dom"
 
-import NavigationBar from "../../../components/Layout/NavigationBar"
-import Footer from "../../../components/Layout/Footer"
+import NavigationBar from "../../../component-library/Layout/NavigationBar"
+import Footer from "../../../component-library/Layout/Footer"
+
 import MaintContentHeader from "../components/MainContentHeader"
 import Account from "../components/Account"
